update update via kaos pushuni command.

+++ universe/Ling/Light_Crud/CrudRequestHandler/LightBaseCrudRequestHandler.php
@@ -54,7 +54,6 @@ class LightBaseCrudRequestHandler implements LightCrudRequestHandlerInterface, L
         }
         $this->checkMicroPermission($pluginContextIdentifier, $table, $action);
 
-
         switch ($action) {
             case "create":
                 $this->executeCreate($pluginContextIdentifier, $table, $params);
@@ -110,6 +109,7 @@ class LightBaseCrudRequestHandler implements LightCrudRequestHandlerInterface, L
 
         $userData = $params['data'];
         $multiplier = $params['multiplier'] ?? null;
+        $useRowRestriction = $params['useRowRestriction'] ?? false;
 
         /**
          * Make sure the user doesn't use the key to do some sql injection.
@@ -141,14 +141,26 @@ class LightBaseCrudRequestHandler implements LightCrudRequestHandlerInterface, L
                 $row = $data;
                 $row[$multiplierColumn] = $val;
                 if (true === $isInsert) {
-                    $db->insert($table, $row);
+                    if (false === $useRowRestriction) {
+                        $db->insert($table, $row);
+                    } else {
+                        $db->pinsert($table, $row);
+                    }
                 } else {
-                    $db->replace($table, $row);
+                    if (false === $useRowRestriction) {
+                        $db->replace($table, $row);
+                    } else {
+                        $db->preplace($table, $row);
+                    }
                 }
             }
 
         } else {
-            $db->insert($table, $data);
+            if (false === $useRowRestriction) {
+                $db->insert($table, $data);
+            } else {
+                $db->pinsert($table, $data);
+            }
         }
     }
 
@@ -182,6 +194,7 @@ class LightBaseCrudRequestHandler implements LightCrudRequestHandlerInterface, L
 
         $userData = $params['data'];
         $userRic = $params['updateRic']; // array of key/value pairs
+        $useRowRestriction = $params['useRowRestriction'] ?? false;
 
 
         /**
@@ -204,7 +217,11 @@ class LightBaseCrudRequestHandler implements LightCrudRequestHandlerInterface, L
          * @var $db LightDatabaseService
          */
         $db = $this->container->get("database");
-        $db->update($table, $data, $ric);
+        if (false === $useRowRestriction) {
+            $db->update($table, $data, $ric);
+        } else {
+            $db->pupdate($table, $data, $ric);
+        }
     }
 
 
@@ -224,6 +241,8 @@ class LightBaseCrudRequestHandler implements LightCrudRequestHandlerInterface, L
          */
         $dbInfoService = $this->container->get("database_info");
         $tableInfo = $dbInfoService->getTableInfo($table);
+        $useRowRestriction = $params['useRowRestriction'] ?? false;
+
 
         /**
          * @var $db LightDatabaseService
@@ -248,8 +267,14 @@ class LightBaseCrudRequestHandler implements LightCrudRequestHandlerInterface, L
 
         $markers = [];
         $sWhere = RicHelper::getWhereByRics($ricStrict, $rics, $markers);
-        $db->delete($table, $sWhere, $markers);
+
+        if (false === $useRowRestriction) {
+            $db->delete($table, $sWhere, $markers);
+        } else {
+            $db->pdelete($table, $sWhere, $markers);
+        }
     }
+
 
 
     /**
